create final CSV2
